+++ test/api_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require File.expand_path(File.join(File.dirname(__FILE__), 'helper'))
 
 class ApiTest < SAPOCI::Connect::TestCase